Return Ocean as a terrain type
It is more coherant that way (ocean is a terrain, after all), and
I still need to properly process the interweaving of terrain types
borders so better to start to do it now.

It's now fixed for Ocean/Plain (beaches), but I'm going to add another
terrain type soon to ensure the overlaying solution works 100% of the
time.