Handle Mountain as a Feature/Special case
It's tricky because mountains are placed over hills, so we need to
distinguish them, hence placing hills as a Feature and mountains as
a Special. But since we need a complete tileset for mountains display,
I have to hack the tile computation algorithm.

Oh well, I may improve that later. For now it just looks better !